+++ updateBloctoExamples.mjs
@@ -1,11 +1,11 @@
-//  updateBloctoExamples.js
+//  updateBloctoExamples
 //
 //  This script is used to update the dependencies of Blocto SDK sample projects.
 //
 //  Usage:
 //  1. Ensure Node.js is installed before running the script.
 //  2. Place the script in the root directory of the sample project.
-//  3. Execute `node updateBloctoExamples.js` in the command line to start the script.
+//  3. Execute `node updateBloctoExamples.mjs` in the command line to start the script.
 //  4. The script will prompt you to choose the type of samples to update (EVM or Aptos) or update all samples.
 //  5. Follow the prompts and inputs; once completed, the script will update the Blocto SDK version for the selected samples.
 //
@@ -19,44 +19,8 @@
 //  - path module for handling file paths.
 //  - prompts module for fetching user input in the command line.
 
-const fs = require("fs").promises;
-const path = require("path");
-const prompts = require("prompts");
-
-async function getDirectories(path) {
-  return fs
-    .readdir(path, { withFileTypes: true })
-    .then((dirents) =>
-      dirents.filter((d) => d.isDirectory()).map((d) => d.name)
-    );
-}
-
-async function updatePackageJSON(folderPath, examples, version) {
-  for (const example of examples) {
-    const packageJSONPath = path.join(folderPath, example, "package.json");
-    try {
-      const packageJSONData = await fs.readFile(packageJSONPath, "utf-8");
-      const packageJSON = JSON.parse(packageJSONData);
-
-      if (packageJSON.dependencies && packageJSON.dependencies["@blocto/sdk"]) {
-        packageJSON.dependencies["@blocto/sdk"] = version;
-        await fs.writeFile(
-          packageJSONPath,
-          JSON.stringify(packageJSON, null, 2)
-        );
-        console.log(
-          `✅ Updated @blocto/sdk version for ${example} to ${version} 🎉`
-        );
-      } else {
-        console.log(`❌ No @blocto/sdk dependency found in ${example}`);
-      }
-    } catch (err) {
-      console.error(
-        `Error updating @blocto/sdk version for ${example}: ${err}`
-      );
-    }
-  }
-}
+import prompts from "prompts";
+import { getDirectories, updatePackageJSON } from "./utils.js";
 
 (async () => {
   const firstResponse = await prompts({
